moved Complex class to outer
